requirements.txt
pytest==3.3.0
requests==2.18.4
faker==0.8.15
allure-pytest==2.3.4b1
requests-toolbelt==0.8.0